03.Django/03.musicapp/pyproject.toml
[tool.poetry]
name = "assignment"
version = "0.1.0"
description = "Pozalabs Backend Engineer Assignment"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.1"
djangorestframework = "^3.13.1"
drf-yasg = "^1.21.3"
django-redis = "^5.2.0"
mypy = "^0.971"
django-stubs = "^1.12.0"
djangorestframework-stubs = "^1.7.0"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
isort = "^5.10.1"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 100
